backend_fastapi/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.4
greenlet==3.3.1
h11==0.16.0
idna==3.11
joblib==1.5.3
nltk==3.9.2
numpy==2.2.6
pandas==2.3.3
passlib==1.7.4
pillow==12.1.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-jose==3.5.0
pytz==2025.2
qrcode==8.2
regex==2026.1.15
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uvicorn==0.40.0
pytesseract==0.3.13
Pillow>=10.0.0
pdfplumber>=0.11.0
